Revert AudioDestinationNode.devicePosition commits
This reverts the following commits:
* 2b2cbc8c ("[Windows] Implementation of 'AudioDestinationNode.devicePosition' attribute")
* 43d41bf6 ("Fix calling of AudioOutputStream::AudioSourceCallback API")
* fd8c95a5 ("Remove unused private field device_position_")

This code was added when Crosswalk was being used as an alternative
implementation for some changes to the WebAudio spec:
* https://github.com/WebAudio/web-audio-api/issues/12
* https://github.com/WebAudio/web-audio-api/pull/754

The spec has since been changed and upstream Chromium is being adapted,
so there is no need to keep these patches in Crosswalk any longer.

Additionally, this code in its present form has caused serious
regressions in WebRTC support (the WebRTC Crosswalk sample did not
produce any sound).

Thanks to Michael Walton <mike@farsouthnet.com> for the investigation
leading to this commit.

BUG=XWALK-7030
BUG=XWALK-7225