Merge 2840: Re-limit the CJK font hack for Android to only when the content language is available
Android FontCache has a hack for CJK, put in 2013 or earlier. This hack
kicked in only when the content language is one of CJK until M53, but
the fix for crbug.com/609043 unintentionally enabled it when the
content language is missing and the system language is one of CJK.

As a quick fix to ease the possible merge to M54, this patch limits the
hack to the same criteria as M53 without reverting crbug.com/609043.

A long term solution is to entirely remove this hack, but it will need
changes to Skia and possibly to fonts.xml as well. This will be worked
out separately.

BUG=652146

Review-Url: https://codereview.chromium.org/2390623002
Cr-Commit-Position: refs/heads/master@{#422424}
(cherry picked from commit 907fee664f47db6be9319c941a7e219ef4052c2b)

Review URL: https://codereview.chromium.org/2396143002 .

Cr-Commit-Position: refs/branch-heads/2840@{#665}
Cr-Branched-From: 1ae106dbab4bddd85132d5b75c670794311f4c57-refs/heads/master@{#414607}